# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List synthetic agents
      parameters:
        - name: district
          in: query
          required: false
          schema: { type: string }
        - name: gender
          in: query
          required: false
          schema: { type: string }
        - name: ageBracket
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
...
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a single agent
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /simulations:
    get:
      operationId: listSimulations
      tags: [simulations]
      summary: List simulation runs
      responses:
        "200":
          description: List of simulations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Simulation"
    post:
      operationId: createSimulation
      tags: [simulations]
      summary: Create a simulation (draft) and get a cost estimate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /simulations/estimate:
    post:
      operationId: estimateSimulation
      tags: [simulations]
      summary: Estimate the cost of a simulation before creating it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateInput"
      responses:
        "200":
          description: Cost estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostEstimate"
  /simulations/{id}:
    get:
      operationId: getSimulation
      tags: [simulations]
      summary: Get a simulation with aggregated results
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Simulation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSimulation
      tags: [simulations]
      summary: Delete a simulation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /simulations/{id}/run:
    post:
      operationId: runSimulation
      tags: [simulations]
      summary: Start the LLM run for a simulation (asynchronous; poll the simulation for progress)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "202":
          description: Run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /simulations/{id}/responses:
    get:
      operationId: listSimulationResponses
      tags: [simulations]
      summary: Per-agent responses for a simulation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationResponse"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Agent:
      type: object
      required:
        - id
        - name
        - age
        - ageBracket
        - gender
        - district
        - lat
        - lng
        - education
        - incomeBracket
        - occupation
        - householdType
        - politicalLeaning
        - partyAffinity
        - turnoutPropensity
        - issueStances
        - consumerStances
        - policyStances
        - mediaDiet
        - values
        - personaSummary
      properties:
        id: { type: integer }
        name: { type: string }
        age: { type: integer }
        ageBracket: { type: string }
        gender: { type: string }
        district: { type: string }
        lat: { type: number }
        lng: { type: number }
        education: { type: string }
        incomeBracket: { type: string }
        occupation: { type: string }
        householdType: { type: string }
        politicalLeaning: { type: integer }
        partyAffinity: { type: string }
        turnoutPropensity: { type: integer }
        issueStances:
          $ref: "#/components/schemas/IssueStances"
        consumerStances:
          $ref: "#/components/schemas/ConsumerStances"
        policyStances:
          $ref: "#/components/schemas/PolicyStances"
        mediaDiet: { type: string }
        values:
          type: array
          items: { type: string }
        personaSummary: { type: string }
    IssueStances:
      type: object
      required: [economy, welfare, security, environment, housing]
      properties:
        economy: { type: integer }
        welfare: { type: integer }
        security: { type: integer }
        environment: { type: integer }
        housing: { type: integer }
    ConsumerStances:
      type: object
      required:
        - priceSensitivity
        - brandLoyalty
        - noveltySeeking
        - ecoConsciousness
        - digitalConsumption
      properties:
        priceSensitivity: { type: integer }
        brandLoyalty: { type: integer }
        noveltySeeking: { type: integer }
        ecoConsciousness: { type: integer }
        digitalConsumption: { type: integer }
    PolicyStances:
      type: object
      required:
        - governmentTrust
        - policyAcceptance
        - taxTolerance
        - regulationPreference
        - publicServiceSatisfaction
      properties:
        governmentTrust: { type: integer }
        policyAcceptance: { type: integer }
        taxTolerance: { type: integer }
        regulationPreference: { type: integer }
        publicServiceSatisfaction: { type: integer }
...
    Simulation:
      type: object
      required:
        - id
        - title
        - audience
        - product
        - policyText
        - model
        - status
        - progress
        - totalAgents
        - costEstimateUsd
        - createdAt
      properties:
        id: { type: integer }
        title: { type: string }
        audience: { type: string }
        product: { type: string }
        policyText: { type: string }
        model: { type: string }
        status: { type: string }
        progress: { type: integer }
        totalAgents: { type: integer }
        costEstimateUsd: { type: number }
        costActualUsd: { type: ["number", "null"] }
        overallSupport: { type: ["number", "null"] }
        supportPct: { type: ["number", "null"] }
        opposePct: { type: ["number", "null"] }
        neutralPct: { type: ["number", "null"] }
        summary: { type: ["string", "null"] }
        createdAt: { type: string }
        completedAt: { type: ["string", "null"] }
    SimulationInput:
      type: object
      required: [title, audience, product, policyText]
      properties:
        title: { type: string, minLength: 1 }
        audience: { type: string }
        product: { type: string }
        policyText: { type: string, minLength: 1 }
        model: { type: string }
    EstimateInput:
      type: object
      required: [model]
      properties:
        model: { type: string }
        totalAgents: { type: integer }
    CostEstimate:
      type: object
      required:
        - model
        - totalAgents
        - inputTokensPerAgent
        - outputTokensPerAgent
        - estimatedCostUsd
        - estimatedLowUsd
        - estimatedHighUsd
        - estimatedSeconds
      properties:
        model: { type: string }
        totalAgents: { type: integer }
        inputTokensPerAgent: { type: integer }
        outputTokensPerAgent: { type: integer }
        estimatedCostUsd: { type: number }
        estimatedLowUsd: { type: number }
        estimatedHighUsd: { type: number }
        estimatedSeconds: { type: integer }
    GroupResult:
      type: object
      required: [key, count, supportPct, opposePct, neutralPct, avgScore]
      properties:
        key: { type: string }
        count: { type: integer }
        supportPct: { type: number }
        opposePct: { type: number }
        neutralPct: { type: number }
        avgScore: { type: number }
    PolicyAxisResult:
      type: object
      required: [axis, label, segments]
      properties:
        axis: { type: string }
        label: { type: string }
        segments:
          type: array
          items: { $ref: "#/components/schemas/GroupResult" }
    SimulationResults:
      type: object
      required: [byDistrict, byAgeBracket, byGender, byLeaning]
      properties:
        byDistrict:
          type: array
          items: { $ref: "#/components/schemas/GroupResult" }
        byAgeBracket:
          type: array
          items: { $ref: "#/components/schemas/GroupResult" }
        byGender:
          type: array
          items: { $ref: "#/components/schemas/GroupResult" }
        byLeaning:
          type: array
          items: { $ref: "#/components/schemas/GroupResult" }
        byPolicyAxis:
          type: array
          description: >-
            Seraph(policy) simulations only — support/acceptance segmented by
            each policy attitude axis (상/중/하). Absent for non-policy sims.
          items: { $ref: "#/components/schemas/PolicyAxisResult" }
    SimulationCalibration:
      type: object
      description: >-
        출력 보정(Lever 2) — 이 시뮬레이션 제품의 과거 검증 이벤트에서 학습한
        평균 편향으로 원시 지지율을 사후 교정한 결과. 이벤트가 부족하면
        applied=false 이고 보정값은 null.
      required: [applied, eventCount, meanBias, shrinkage, events]
      properties:
        applied: { type: boolean }
        eventCount: { type: integer }
        meanBias: { type: number }
        shrinkage: { type: number }
        calibratedSupportPct: { type: ["number", "null"] }
        calibratedOpposePct: { type: ["number", "null"] }
        calibratedNeutralPct: { type: ["number", "null"] }
        events:
          type: array
          description: 이 보정에 사용된 과거 검증 이벤트 목록 (어떤 검증이 쓰였는지 명시).
          items:
            type: object
            required: [id, title, eventType, targetDate, actualValue, rawPrediction, bias]
            properties:
              id: { type: integer }
              title: { type: string }
              eventType: { type: string }
              targetDate: { type: string }
              actualValue: { type: number }
              rawPrediction: { type: number }
              bias: { type: number }
    SimulationDetail:
      type: object
      required: [simulation, results]
      properties:
        simulation:
          $ref: "#/components/schemas/Simulation"
        results:
          $ref: "#/components/schemas/SimulationResults"
        calibration:
          $ref: "#/components/schemas/SimulationCalibration"
    SimulationResponse:
      type: object
      required:
        - id
        - simulationId
        - agentId
        - agentName
        - district
        - ageBracket
        - gender
        - stance
        - score
        - confidence
        - reasoning
      properties:
        id: { type: integer }
        simulationId: { type: integer }
        agentId: { type: integer }
        agentName: { type: string }
        district: { type: string }
        ageBracket: { type: string }
        gender: { type: string }
        stance: { type: string }
        score: { type: integer }
        confidence: { type: integer }
        reasoning: { type: string }
...